# docs/api/openapi-assessment-orchestrator.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Every non-2xx body. One shape, so a client has one error path.",
        "properties": {
          "code": {
            "default": "",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "assessment-orchestrator \u2014 the loop, the posterior, selection and stopping.\n\nWHAT THIS SERVICE DECIDES, AND WHAT IT DELEGATES\n\nIt owns exactly two decisions: WHICH competency to probe next, and WHEN one is finished.\nWhich item measures a competency belongs to the Picking Agent; how a response is graded\nbelongs to the grader service; what a response implies about neighbouring nodes belongs to\nthe graph service. Keeping those apart is what lets a modality be added without touching\nthe loop.\n\nA model picks an item from a shortlist the engine has already ranked, and interprets code.\nIt does not grade multiple choice, does not write an estimate, does not choose which\ncompetency to probe, and cannot stop an assessment. Every number in a report is computed by\ndeterministic code from inputs recorded beside it.\n\nSTATELESS ENGINE, STATEFUL SERVICE\n\nThe `Orchestrator` takes a state and returns one; it holds nothing. This service holds the\nstates, in memory, with the monolith's retention rules \u2014 see `sessions.py` for why that is\ndeliberate and what it costs.\n\nTHE CANDIDATE BOUNDARY IS HERE\n\nResponses on the assessment routes carry no answer key, no hidden test and no grading\ndetail. That is enforced by which types they return, not by a client choosing what to\nrender \u2014 see `presentation.py`. The instrumented view lives behind\n`AUTHOR_DIAGNOSTICS_ENABLED` and is refused, not thinned, when off.\n",
...
    "version": "dev"
...
  "openapi": "3.1.0",
  "paths": {
    "/assessments": {
      "post": {
        "description": "The bank id travels with the session from here on, not with each request: a client that began against one bank must not be able to answer it against another, and re-sending the id every call would make that possible by omission.\n\nThe bank VERSION is pinned too. Replacing a bank mid-session would otherwise change the item pool underneath a live candidate.",
        "operationId": "create_assessment_assessments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "no such assessment"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the presented item changed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the answer does not fit the item"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "a dependency is unreachable"
          }
        },
        "summary": "Begin an assessment over several competencies",
        "tags": [
          "assessment"
        ]
      }
    },
    "/assessments/{assessment_id}": {
      "delete": {
        "operationId": "delete_assessment_assessments__assessment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Assessment Assessments  Assessment Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "no such assessment"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the presented item changed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the answer does not fit the item"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "a dependency is unreachable"
          }
        },
        "summary": "Discard an assessment",
        "tags": [
          "assessment"
        ]
      },
      "get": {
        "operationId": "get_assessment_assessments__assessment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "no such assessment"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the presented item changed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the answer does not fit the item"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "a dependency is unreachable"
          }
        },
        "summary": "Current state, with the presented item or the report",
        "tags": [
          "assessment"
        ]
      }
    },
    "/assessments/{assessment_id}/diagnostics": {
      "get": {
        "description": "Refused unless `AUTHOR_DIAGNOSTICS_ENABLED`. It carries the posterior, the shortlist and which item the engine would have picked \u2014 enough to reverse-engineer difficulty, and enough for a candidate to tell how they are doing while they are still being measured.",
        "operationId": "diagnostics_assessments__assessment_id__diagnostics_get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "no such assessment"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the presented item changed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the answer does not fit the item"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "a dependency is unreachable"
          }
        },
        "summary": "Posteriors, selection reasoning and graph state. NOT candidate-safe.",
        "tags": [
          "author"
        ]
      }
    },
    "/assessments/{assessment_id}/next": {
      "get": {
        "description": "A convenience over `GET /assessments/{id}`, for a client that only wants the question. 409 once the assessment has stopped \u2014 an empty 200 would read as 'nothing to do yet' rather than 'this is over'.",
        "operationId": "next_item_assessments__assessment_id__next_get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentingDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "no such assessment"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the assessment has stopped"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the answer does not fit the item"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "a dependency is unreachable"
          }
        },
        "summary": "The next item to present",
        "tags": [
          "assessment"
        ]
      }
    },
...
    "/banks": {
      "get": {
        "operationId": "list_banks_banks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BankSummary"
                  },
                  "title": "Response List Banks Banks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "no such assessment"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the presented item changed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the answer does not fit the item"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "a dependency is unreachable"
          }
        },
        "summary": "Registered banks, proxied so a frontend has one base URL",
        "tags": [
          "banks"
        ]
      }
    },
...
  }
}